Move `PaginationMetadata` into `common.controller`
